Move do_clear(), do_execproc(), do_set_password(),do_refresh_metadata_cache(), do_show_db(),do_change_db(), do_whoami() functions into do_utils file from easyaccess
